+++ src/Ninject.Test/Integration/EnumerableDependenciesTests/Fakes/RequestsListWithConcreteClass.cs
@@ -1,0 +1,14 @@
+﻿namespace Ninject.Tests.Integration.EnumerableDependenciesTests.Fakes
+{
+    using System.Collections.Generic;
+
+    public class RequestsListWithConcreteClass
+    {
+        public RequestsListWithConcreteClass(IList<ChildA> children)
+        {
+            this.Children = children;
+        }
+
+        public IList<ChildA> Children { get; private set; }
+    }
+}